chatbot/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
Bottleneck==1.6.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dataclasses-json==0.6.7
datasets==4.3.0
dill==0.4.0
docstring_parser==0.17.0
filelock==3.20.0
fpdf==1.7.2
frozenlist==1.8.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-auth==2.42.0
google-cloud-aiplatform==1.123.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.15.0
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.47.0
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.0.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
json_repair==0.30.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79
langchain-experimental==0.3.4
langchain-google-vertexai==2.1.2
langchain-neo4j==0.4.0
langchain-text-splitters==0.3.11
langsmith==0.4.38
MarkupSafe==3.0.3
marshmallow==3.26.1
multidict==6.7.0
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==2.10.0
neo4j==5.28.2
neo4j-graphrag==1.5.0
numexpr==2.14.1
numpy==2.3.4
orjson==3.11.4
packaging==24.2
pandas==2.3.3
pillow==12.0.0
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.9.1
pydantic-settings==2.11.0
pydantic_core==2.23.3
pydeck==0.9.1
Pygments==2.19.2
pypdf==4.3.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.28.0
rsa==4.9.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
streamlit==1.51.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typer-slim==0.20.0
types-PyYAML==6.0.12.20250915
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uv==0.9.6
validators==0.35.0
watchdog==6.0.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
